+++ lib/reservation_monitor.py
@@ -64,7 +64,6 @@ class ReservationMonitor:
             time.sleep(0.05)
             # Lock so all processes are stopped sequentially
             with self.lock:
-                WebDriver.reset_temp_dir()
                 self._stop_monitoring()
 
     def _monitor(self) -> None:
@@ -94,7 +93,7 @@ class ReservationMonitor:
         Check for reservation changes and lower fares. Returns true if future checks should not be
         performed (e.g. no more flights are scheduled to check in).
         """
-        reservation = {"confirmationNumber": self.config.confirmation_number}
+        reservation = {"record_locator": self.config.confirmation_number}
 
         # Ensure there are valid headers
         try:
@@ -116,7 +115,7 @@ class ReservationMonitor:
 
     def _schedule_reservations(self, reservations: list[dict[str, Any]]) -> None:
         logger.debug("Scheduling flight check-ins for %d reservations", len(reservations))
-        confirmation_numbers = [reservation["confirmationNumber"] for reservation in reservations]
+        confirmation_numbers = [reservation["record_locator"] for reservation in reservations]
         self.checkin_scheduler.process_reservations(confirmation_numbers)
 
     def _check_flight_fares(self) -> None:
@@ -251,7 +250,6 @@ class AccountMonitor(ReservationMonitor):
                             err.status_code,
                         )
                         logger.debug("Waiting for %d seconds before retrying", RETRY_WAIT_SECONDS)
-                        webdriver.reset_temp_dir()
                         time.sleep(RETRY_WAIT_SECONDS)
                     else:
                         logger.debug(