+++ tests/integration/test_monitoring_and_scheduling.py
@@ -161,13 +161,7 @@ def test_account_schedules_new_flights(requests_mock: RequestMocker, mocker: Moc
             '"customers.userInformation.lastName": "Gump"}'
         )
     }
-
-    trips_response = {
-        "body": (
-            '{"upcomingTripsPage": [{"tripType": "FLIGHT", "confirmationNumber": "TEST"}, '
-            '{"tripType": "CAR"}]}'
-        )
-    }
+    trips_response = {"body": ('{"data": [{"record_locator": "TEST"}]}')}
 
     login_attempts = 0