+++ tests/unit/test_webdriver.py
@@ -6,7 +6,13 @@ import pytest
 from pytest_mock import MockerFixture
 
 from lib.utils import DriverTimeoutError, LoginError
-from lib.webdriver import HEADERS_URL, INVALID_CREDENTIALS_CODE, LOGIN_URL, TRIPS_URL, WebDriver
+from lib.webdriver import (
+    INVALID_CREDENTIALS_CODE,
+    MOBILE_HEADERS_URL,
+    SUCCESSFUL_LOGIN_URL,
+    TRIPS_URL,
+    WebDriver,
+)
 
 
 @pytest.fixture(autouse=True)
@@ -17,40 +23,6 @@ def mock_chrome(mocker: MockerFixture) -> mock.Mock:
 @pytest.fixture
 def mock_account_monitor(mocker: MockerFixture) -> mock.Mock:
     return mocker.patch("lib.reservation_monitor.AccountMonitor")
-
-
-# Tests for the WebDriver's class methods
-
-
-def test_get_temp_dir_returns_temp_directory() -> None:
-    WebDriver._temp_dir = "/mock/temp/dir"
-    assert WebDriver.get_temp_dir() == "/mock/temp/dir"
-
-
-def test_reset_temp_dir_creates_directory_on_first_use(mocker: MockerFixture) -> None:
-    mocker.patch("tempfile.mkdtemp", return_value="/mock/temp/new-dir")
-    WebDriver._temp_dir = None
-    assert WebDriver.get_temp_dir() == "/mock/temp/new-dir"
-
-
-def test_reset_temp_dir_removes_existing_directory(mocker: MockerFixture) -> None:
-    mock_rmtree = mocker.patch("shutil.rmtree")
-    mock_exists = mocker.patch("os.path.exists", return_value=True)
-    WebDriver._temp_dir = "/mock/temp/dir"
-    WebDriver.reset_temp_dir()
-
-    mock_exists.assert_called_once_with("/mock/temp/dir")
-    mock_rmtree.assert_called_once_with("/mock/temp/dir")
-
-
-def test_reset_temp_dir_ignores_if_temp_dir_is_not_set(mocker: MockerFixture) -> None:
-    mock_rmtree = mocker.patch("shutil.rmtree")
-    mock_exists = mocker.patch("os.path.exists")
-    WebDriver._temp_dir = None
-    WebDriver.reset_temp_dir()
-
-    mock_exists.assert_not_called()
-    mock_rmtree.assert_not_called()
 
 
 class TestWebDriver:
@@ -91,7 +63,6 @@ class TestWebDriver:
         self, mocker: MockerFixture, mock_chrome: mock.Mock, mock_account_monitor: mock.Mock
     ) -> None:
         mocker.patch("time.sleep")
-        mocker.patch("lib.webdriver.seleniumbase_actions.wait_for_element_not_visible")
         mocker.patch.object(WebDriver, "_get_driver", return_value=mock_chrome)
         mock_wait_for_attribute = mocker.patch.object(self.driver, "_wait_for_attribute")
         mock_wait_for_login = mocker.patch.object(WebDriver, "_wait_for_login")
@@ -104,6 +75,8 @@ class TestWebDriver:
         mock_wait_for_attribute.assert_called_once()
         mock_wait_for_login.assert_called_once()
         mock_chrome.add_cdp_listener.assert_called_once()
+        # Ensure the driver navigates to the normal website to fetch reservations
+        mock_chrome.get.assert_called_once()
         mock_chrome.quit.assert_called_once()
 
     def test_get_driver_returns_a_webdriver_with_one_request(self, mock_chrome: mock.Mock) -> None:
@@ -132,7 +105,7 @@ class TestWebDriver:
 
     def test_headers_listener_sets_headers_when_correct_url(self, mocker: MockerFixture) -> None:
         mocker.patch.object(self.driver, "_get_needed_headers", return_value={"test": "headers"})
-        data = {"params": {"request": {"url": HEADERS_URL, "headers": {}}}}
+        data = {"params": {"request": {"url": MOBILE_HEADERS_URL, "headers": {}}}}
 
         self.driver._headers_listener(data)
 
@@ -147,7 +120,12 @@ class TestWebDriver:
         assert self.driver.checkin_scheduler.headers == {}
 
     def test_login_listener_sets_login_information(self) -> None:
-        data = {"params": {"response": {"url": LOGIN_URL, "status": 200}, "requestId": "test_id"}}
+        data = {
+            "params": {
+                "response": {"url": SUCCESSFUL_LOGIN_URL, "status": 200},
+                "requestId": "test_id",
+            }
+        }
         self.driver._login_listener(data)
 
         assert self.driver.login_status_code == 200
@@ -224,39 +202,41 @@ class TestWebDriver:
     ) -> None:
         mocker.patch("seleniumbase.fixtures.page_actions.wait_for_element_not_visible")
         mocker.patch.object(mock_chrome, "is_element_visible", return_value=False)
+
         self.driver._click_login_button(mock_chrome)
         mock_chrome.click.assert_not_called()
 
-    def test_click_login_button_does_not_click_when_popup_appears(
+    def test_click_login_button_does_not_click_after_failed_login(
         self, mocker: MockerFixture, mock_chrome: mock.Mock
     ) -> None:
+        # Simulate an error message being visible after logging in
         mock_wait_for_element = mocker.patch(
             "seleniumbase.fixtures.page_actions.wait_for_element_not_visible"
         )
         mocker.patch.object(mock_chrome, "is_element_visible", return_value=True)
+
         self.driver._click_login_button(mock_chrome)
-        mock_wait_for_element.assert_called_once()
+        mock_wait_for_element.assert_not_called()
+        mock_chrome.click.assert_not_called()
 
     def test_click_login_button_clicks_when_form_fails_to_submit(
         self, mocker: MockerFixture, mock_chrome: mock.Mock
     ) -> None:
         mocker.patch(
-            "seleniumbase.fixtures.page_actions.wait_for_element_not_visible",
-            side_effect=[None, Exception],
+            "seleniumbase.fixtures.page_actions.wait_for_element_not_visible", side_effect=Exception
         )
         mocker.patch.object(mock_chrome, "is_element_visible", return_value=False)
+
         self.driver._click_login_button(mock_chrome)
         mock_chrome.click.assert_called_once()
 
-    def test_fetch_reservations_fetches_only_flight_reservations(
-        self, mocker: MockerFixture
-    ) -> None:
-        trips_response = {"upcomingTripsPage": [{"tripType": "FLIGHT"}, {"tripType": "CAR"}]}
+    def test_fetch_reservations_fetches_flight_reservations(self, mocker: MockerFixture) -> None:
+        trips_response = {"data": ["flight1", "flight2"]}
 
         mocker.patch.object(WebDriver, "_wait_for_attribute")
         mocker.patch.object(WebDriver, "_get_response_body", return_value=trips_response)
 
-        assert self.driver._fetch_reservations(None) == [{"tripType": "FLIGHT"}]
+        assert self.driver._fetch_reservations(None) == ["flight1", "flight2"]
 
     def test_get_response_body_loads_body_from_response(self, mock_chrome: mock.Mock) -> None:
         mock_chrome.execute_cdp_cmd.return_value = {"body": '{"response": "body"}'}
@@ -313,6 +293,22 @@ class TestWebDriver:
         )
 
         assert mock_account_monitor.first_name == "John"
+        assert mock_account_monitor.last_name == "Doe"
+
+    def test_set_account_name_sets_the_preferred_first_name_when_available(
+        self, mock_account_monitor: mock.Mock
+    ) -> None:
+        mock_account_monitor.first_name = None
+        self.driver._set_account_name(
+            mock_account_monitor,
+            {
+                "customers.userInformation.firstName": "John",
+                "customers.userInformation.lastName": "Doe",
+                "customers.userInformation.preferredName": "Johnny",
+            },
+        )
+
+        assert mock_account_monitor.first_name == "Johnny"
         assert mock_account_monitor.last_name == "Doe"
 
     def test_quit_driver_cleans_up_webdriver(